+++ arklight/ir/__init__.py
@@ -1,7 +1,7 @@
 from arklight.ir.normalize import normalize_ark_ast, normalize_node
 from arklight.ir.validate import ValidationError, validate_ark_ast
 from arklight.ir.build import WebsiteIR, IRNode, IRPage, build_website_ir
-from arklight.ir.schema import SCHEMA, NodeSpec, TEXT_ONLY_TYPES
+from arklight.ir.schema import SCHEMA, NodeSpec, TEXT_ONLY_TYPES, KNOWN_BEHAVIORS
 
 __all__ = [
     "normalize_ark_ast",
@@ -15,4 +15,5 @@ __all__ = [
     "SCHEMA",
     "NodeSpec",
     "TEXT_ONLY_TYPES",
+    "KNOWN_BEHAVIORS",
 ]